Add first structure
